pyproject.toml
[build-system]
requires = ["uv_build>=0.8.22"]
build-backend = "uv_build"

[project]
name = "aind-behavior-implementation-example"
description = "A library that defines AIND data schema for a behavior experiment."
authors = [ {name = "RoboDoig", email = "erskine94@gmail.com"}]
license = "MIT"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Operating System :: Microsoft :: Windows",
]
version = "0.0.0rc0"
readme = {file = "README.md", content-type = "text/markdown"}

dependencies = [
  "aind_behavior_services>=0.12, <0.13",
  "pydantic-settings"
]

[project.urls]
Documentation = "https://allenneuraldynamics.github.io/Aind.Behavior.ImplementationExample/"
Repository = "https://github.com/AllenNeuralDynamics/Aind.Behavior.ImplementationExample/"
Issues = "https://github.com/AllenNeuralDynamics/Aind.Behavior.ImplementationExample/issues"
Changelog = "https://github.com/AllenNeuralDynamics/Aind.Behavior.ImplementationExample/releases"

[project.optional-dependencies]

data = ["contraqctor>=0.5.3, <0.6.0"]

launcher = [
    "aind-clabe[aind-services]>=0.9.1",
    "aind-data-schema>=2.2",
    "aind-behavior-implementation-example[data]",
    ]

[dependency-groups]

dev = [
    "aind-behavior-implementation-example[launcher]",
    "aind-behavior-implementation-example[data]",
    "ruff",
    "codespell",
]

docs = [
    'Sphinx',
    'furo',
    'sphinx-jinja',
    'autodoc_pydantic[erdantic]',
    'myst_parser',
    'sphinx-jsonschema',
    "aind-behavior-implementation-example[data]",  # Necessary for the data contract rendering
]

[project.scripts]
implementation-example = "aind_behavior_implementation_example.cli:main"

[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.ruff.lint]
extend-select = ['Q', 'RUF100', 'C90', 'I']
extend-ignore = []
mccabe = { max-complexity = 14 }
pydocstyle = { convention = 'google' }

[tool.codespell]
skip = '.git,*.pdf,*.svg,./bonsai,*.bonsai,./docs/_build,uv.lock,*Generated.cs'
ignore-words-list = 'nd'

[tool.uv]
default-groups = ['dev']
required-version = '>=0.8.4'
